Follow-up to r155: Migrated DATABASE_ODBC_DRIVER, DATABASE_ODBC_DSN and DATABASE_ODBC_EXTRA_PARAMS settings to DATABASE_OPTIONS.
This is a backward-incompatible change for setups that were using al least one
of these three settings.

The equivalente DATABASE_OPTIONS keys are 'driver, 'dsn' and 'extra_params'
respectively.

The DATABASE_COLLATION setting will be subject to the same migration
(to DATABASE_OPTIONS['collation']) soon (it needs some additional
re-factoring), for now it raises a DeprecationWarning.


git-svn-id: http://django-pyodbc.googlecode.com/svn/trunk@157 d8ed95fc-f742-0410-923b-f1a0a12816d3