Updating pypi/pip version to non-backwards compatible `2.0.0a1`.
Django 2.0 is a major release that [only supports python 3.4+][1] and doesn't
support python 2.x. Tracking with this major Django change, the django-pyodbc
releases greater than major version 2 are also jettisoning support for python
2.x, having run [2to3][4], and doing backwards-incompatible edits to the library to simplify
maintenance tasks. This is a compatibility-breaking change and we are changing
the MAJOR release number in accordance with [semver][3]. Since the testing of
this release hasn't been rigorous, we are initially releasing it on a
pre-release basis until testing coverage can be updated and automated. This will
not break folks who are installing django-pyodbc via pip, because pip [does not
install pre-release versions of libraries][2] without the `--pre` switch or
using a specific release like:

    pip install django-pyodbc==2.0.0a1

Going forward, fixes and updates for Django 2 should be committed to branch
`master` and any fixes needed for Django 1.x releases should be committed to
branch `django-1`.

  [1]: https://docs.djangoproject.com/en/2.0/releases/2.0/#python-compatibility
  [2]: https://pip.pypa.io/en/stable/reference/pip_install/#pre-release-versions
  [3]: https://semver.org/
  [4]: https://docs.python.org/2/library/2to3.html